--- a/src/ducks/SANCharts/utils.js
+++ b/src/ducks/SANCharts/utils.js
@@ -104,6 +104,7 @@ export const Metrics = {
   },
   transactionVolume: {
     alias: 'transaction_volume',
+    dataKey: 'transaction_volume',
     category: 'On-chain',
     node: Bar,
     group: 'Token Flows/Movement/Activity',