… src/ducks/GetTimeSeries/GetTimeSeries.spec.js
… src/ducks/GetTimeSeries/__snapshots__/GetTimeSeries.spec.js.snap
… src/ducks/GetTimeSeries/__snapshots__/epics.spec.js.snap
… src/ducks/GetTimeSeries/epics.spec.js
… src/pages/Chart/index.js
… src/pages/Detailed/MobileDetailedPage.js
… src/pages/Trends/TrendsExplorePage.js