### Summary
- Improvements of the `GetTimeSeries` component;
- Using `getMetric` endpoint for the `transaction_volume` metric. More metrics yet to come.

### What's changed?
- Required metrics are now passed as the array objects via [`metrics` prop](https://github.com/santiment/app/pull/554/files#diff-34bac4f2ff6d26d903ac69c7aa58c12cR31). Every object should contain [`name` property](https://github.com/santiment/app/pull/554/files#diff-34bac4f2ff6d26d903ac69c7aa58c12cR33) that corresponds to a metric's name.
- No more `settings` since it was never used.

### Benchmarks
Benchmarks were made on the laptop in the `development` build. The results are the average of `10` runs. 
During `Preparing request` benchmark, [this section](https://github.com/santiment/app/blob/dc625cf979ed7606b3e38459c2b05d9088639f4d/src/ducks/GetTimeSeries/epics.js#L27-L75) was measured on the current branch, and [this section](https://github.com/santiment/app/blob/master/src/ducks/GetTimeSeries/epics.js#L70-L114) on the master branch.
During `Forming response` benchmark, [this section](https://github.com/santiment/app/blob/dc625cf979ed7606b3e38459c2b05d9088639f4d/src/ducks/GetTimeSeries/epics.js#L78-L83) was measured on the current branch, and [this section](https://github.com/santiment/app/blob/master/src/ducks/GetTimeSeries/epics.js#L117-L122) on the master branch.

#### First time after emptying cache and hard-reload
Old:
- Preparing request: `67,9ms`;
- Forming response: `2,1ms`;

New: 
- Preparing request: `33,5ms`;
- Forming response: `1,1ms`;

#### Following interactions
Old:
- Preparing request: `3,2ms`;
- Forming response: `1,3ms`;

New: 
- Preparing request: `2,6ms`;
- Forming response: `0,76ms`;